+++ app/javascript/engine/actions/RoutCheckAction.ts
@@ -1,6 +1,6 @@
 import { Coordinate } from "../../utilities/commonTypes";
-import { failRed, formatCoordinate, formatDieResult, formatNation, formatTarget, passBlue } from "../../utilities/graphics";
-import { baseMorale, otherPlayer } from "../../utilities/utilities";
+import { failRed, formatCoordinate, formatDieResult, formatNation, formatTarget, passBlue, passGreen } from "../../utilities/graphics";
+import { baseMorale, otherPlayer, playerForNation } from "../../utilities/utilities";
 import Game from "../Game";
 import { GameActionData, GameActionDiceResult, GameActionRoutData, GameActionUnit } from "../GameAction";
 import Unit from "../Unit";
@@ -26,7 +26,6 @@ export default class RoutCheckAction extends BaseAction {
   get type(): string { return "rout_check" }
 
   get htmlValue(): string {
-    // const unit = this.game.findUnitById(this.target.id) as Unit
     const nation = formatNation(this.game, otherPlayer(this.player))
     const loc = formatCoordinate(new Coordinate(this.target.x, this.target.y))
     const check = baseMorale + this.routCheckMods.mod - 2
@@ -35,6 +34,9 @@ export default class RoutCheckAction extends BaseAction {
       `rout morale check at ${loc}: target ${formatTarget(check)}, rolled ${formatDieResult(roll)}, `
     if (roll.result < check) {
       rc += `<span style="color: ${failRed()};">failed</span>, unit routs`
+    } else if (this.routCheckMods.rally_mod !== undefined && roll.result >= check + 12 - this.routCheckMods.rally_mod) {
+      // TODO: see below
+      rc += `<span style="color: ${passGreen()};">critical success</span>, unit rallies`
     } else {
       rc += `<span style="color: ${passBlue()};">passed</span>, no effect`
     }
@@ -54,6 +56,15 @@ export default class RoutCheckAction extends BaseAction {
     if (roll.result < check) {
       this.game.routNeeded.push({ unit, loc })
       this.game.addActionAnimations([{ loc, type: "rout" }])
+    } else if (this.routCheckMods.rally_mod !== undefined && roll.result >= check + 12 - this.routCheckMods.rally_mod) {
+      // TODO: check for existing data; won't break existing moves, theoretically
+      // could pull this after all games before id=69 deleted, if ever
+      unit.resetStatus()
+      this.game.addActionAnimations([{ loc, type: "rally" }])
+      const vp = this.map.victoryAt(loc)
+      if (vp && vp !== playerForNation(unit, this.game) && !this.map.contactAt(loc)) {
+        this.map.toggleVP(loc)
+      }
     } else {
       this.game.addActionAnimations([{ loc, type: "norout" }])
     }